Backend/requirements.txt
langchain 
langchain-community 
langchain-core 
langchain-huggingface 
huggingface_hub 
datasets 
faiss-cpu 
moviepy 
pydub 
ffmpeg-python 
fastapi 
uvicorn 
python-multipart 
pydantic 
requests 
tqdm 
numpy 
pandas 
scikit-learn 
matplotlib 
jupyter 
ipywidgets 
ollama
rank-bm25
groq
yt-dlp
youtube-transcript-api
fastembed
google-cloud-storage